스레드 교육 예제 완성

+++ thisisjava/src/ch17/sec13/ParallelExample.java
@@ -10,7 +10,7 @@ public class ParallelExample {
 		Random random = new Random();
 
 		List<Integer> scores = new ArrayList< >();
-		for(int i=0; i<100000000; i++) {
+		for(long i=0; i<100000000; i++) {
 			scores.add(random.nextInt(101));
 		}
